Correct filesystem stream integer widths - Preserve full write counts, positions and file sizes - Match current Linux filesystem call paths

+++ public/filesystem.h
@@ -416,17 +416,17 @@ public:
 	// IBaseFileSystem
 	//--------------------------------------------------------
 	virtual int				Read(void* pOutput, int size, FileHandle_t file) = 0;
-	virtual int				Write(void const* pInput, int size, FileHandle_t file) = 0;
+	virtual int				Write(void const* pInput, uint64 size, FileHandle_t file) = 0;
 
 	// if pathID is NULL, all paths will be searched for the file
 	virtual FileHandle_t	Open(const char *pFileName, const char *pOptions, const char *pathID = 0) = 0;
 	virtual void			Close(FileHandle_t file) = 0;
 
 
-	virtual void			Seek(FileHandle_t file, int pos, FileSystemSeek_t seekType) = 0;
-	virtual unsigned int	Tell(FileHandle_t file) = 0;
-	virtual unsigned int	Size(FileHandle_t file) = 0;
-	virtual unsigned int	Size(const char *pFileName, const char *pPathID = 0) = 0;
+	virtual void			Seek(FileHandle_t file, int64 pos, FileSystemSeek_t seekType) = 0;
+	virtual int64			Tell(FileHandle_t file) = 0;
+	virtual int64			Size(FileHandle_t file) = 0;
+	virtual int64			Size(const char *pFileName, const char *pPathID = 0) = 0;
 
 	virtual void			Flush(FileHandle_t file) = 0;
 	virtual bool			Precache(const char *pFileName, const char *pPathID = 0) = 0;
